+++ src/commands/basemaps-github/create-pr.ts
@@ -15,7 +15,15 @@ export const ValidSourceBuckets: Set<string> = new Set(['nz-imagery', 'linz-imag
 
 export const LinzBasemapsSourceCollectionRel = 'linz_basemaps:source_collection';
 
-async function parseTargetInfo(
+function assertValidBucket(bucket: string, validBuckets: Set<string>): void {
+  // Validate the target information
+  logger.info({ bucket }, 'CreatePR: Valid the target s3 bucket');
+  if (!validBuckets.has(bucket)) {
+    throw new Error(`Invalid s3 bucket ${bucket} from the target.`);
+  }
+}
+
+async function parseRasterTargetInfo(
   target: string,
   individual: boolean,
 ): Promise<{ name: string; title: string; epsg: EpsgCode; region: string | undefined }> {
@@ -26,11 +34,7 @@ async function parseTargetInfo(
   const epsg = Epsg.tryGet(Number(splits[1]));
   const name = splits[2];
 
-  //Validate the target information
-  logger.info({ bucket }, 'CreatePR: Valid the target s3 bucket');
-  if (bucket == null || !ValidTargetBuckets.has(bucket)) {
-    throw new Error(`Invalid s3 bucket ${bucket} from the target ${target}.`);
-  }
+  assertValidBucket(bucket, ValidTargetBuckets);
 
   if (epsg == null || name == null) throw new Error(`Invalid target ${target} to parse the epsg and imagery name.`);
   const collectionPath = fsa.join(target, 'collection.json');
@@ -44,10 +48,8 @@ async function parseTargetInfo(
   if (source == null) throw new Error(`Failed to get source url from collection.json.`);
   const sourceUrl = new URL(source);
   const sourceBucket = sourceUrl.hostname;
-  logger.info({ bucket: sourceBucket }, 'CreatePR: Validate the source s3 bucket');
-  if (sourceBucket == null || !ValidSourceBuckets.has(sourceBucket)) {
-    throw new Error(`Invalid s3 bucket ${sourceBucket} from the source ${sourceUrl}.`);
-  }
+  assertValidBucket(sourceBucket, ValidSourceBuckets);
+
   // Try to get the region for individual layers
   let region;
   if (individual) {
@@ -61,6 +63,28 @@ async function parseTargetInfo(
   }
 
   return { name: standardizeLayerName(name), epsg: epsg.code, title, region };
+}
+
+/**
+ * Pass vector target location with the following format
+ * s3://linz-basemaps-staging/vector/3857/53382-nz-roads-addressing/01HSF04SG9M1P3V667A4NZ1MN8/53382-nz-roads-addressing.tar.co
+ * s3://linz-basemaps-staging/vector/3857/topographic/01HSF04SG9M1P3V667A4NZ1MN8/topographic.tar.co
+ */
+async function parseVectorTargetInfo(target: string): Promise<{ name: string; title: string; epsg: EpsgCode }> {
+  logger.info({ target }, 'CreatePR: Get the layer information from target');
+  const url = new URL(target);
+  const bucket = url.hostname;
+  const splits = url.pathname.split('/');
+  const epsg = Epsg.tryGet(Number(splits[2]));
+  const name = splits[3];
+
+  assertValidBucket(bucket, validTargetBuckets);
+
+  if (epsg == null || name == null) throw new Error(`Invalid target ${target} to parse the epsg and imagery name.`);
+  if (epsg !== Epsg.Google) throw new Error(`Unsupported epsg code ${epsg.code} for vector map.`);
+  // Try to get the region for individual layers
+
+  return { name: name, epsg: epsg.code, title: name };
 }
 
 export const CommandCreatePRArgs = {
@@ -94,6 +118,11 @@ export const CommandCreatePRArgs = {
     long: 'vector',
     description: 'Import layer into vector config in basemaps.',
   }),
+  ticket: option({
+    type: optional(string),
+    long: 'ticket',
+    description: 'Associated JIRA ticket e.g. AIP-74',
+  }),
 };
 
 export const basemapsCreatePullRequest = command({
@@ -115,12 +144,15 @@ export const basemapsCreatePullRequest = command({
     const layer: ConfigLayer = { name: '', title: '' };
     let region;
     if (args.vector) {
-      layer.name = 'topographic';
-      layer.title = 'Topographic';
-      layer[3857] = targets[0];
+      for (const target of targets) {
+        const info = await parseVectorTargetInfo(target);
+        layer.name = info.name;
+        layer.title = info.title;
+        layer[info.epsg] = target;
+      }
     } else {
       for (const target of targets) {
-        const info = await parseTargetInfo(target, args.individual);
+        const info = await parseRasterTargetInfo(target, args.individual);
         layer.name = info.name;
         layer.title = info.title;
         layer[info.epsg] = target;
@@ -131,8 +163,8 @@ export const basemapsCreatePullRequest = command({
 
     if (layer.name === '' || layer.title === '') throw new Error('Failed to find the imagery name or title.');
 
-    const git = new MakeCogGithub(layer.name, args.repository);
-    if (args.vector) await git.updateVectorTileSet('topographic', layer);
+    const git = new MakeCogGithub(layer.name, args.repository, args.ticket);
+    if (args.vector) await git.updateVectorTileSet(layer.name, layer);
     else await git.updateRasterTileSet('aerial', layer, category, region);
   },
 });