+++ src/commands/path/path.generate.ts
@@ -1,6 +1,6 @@
 import { Epsg } from '@basemaps/geo';
 import { fsa } from '@chunkd/fs';
-import { CogTiff } from '@cogeotiff/core';
+import { Tiff } from '@cogeotiff/core';
 import { command, option, positional, string } from 'cmd-ts';
 import { StacCollection, StacItem } from 'stac-ts';
 
@@ -131,8 +131,8 @@ export function formatName(region: string, geographicDescription?: string): stri
 
 export function formatDate(collection: StacCollection): string {
   const interval = collection.extent?.temporal?.interval?.[0];
-  const startYear = interval[0]?.slice(0, 4);
-  const endYear = interval[1]?.slice(0, 4);
+  const startYear = getPacificAucklandYear(interval[0]);
+  const endYear = getPacificAucklandYear(interval[1]);
 
   if (startYear == null || endYear == null) {
     throw new Error(`Missing datetime in interval: ${interval}`);
@@ -141,6 +141,29 @@ export function formatDate(collection: StacCollection): string {
     return startYear;
   }
   return `${startYear}-${endYear}`;
+}
+
+/**
+ * Convert time zone-aware date/time string to Pacific/Auckland time zone string
+ *
+ * We can't convert the time zone of a `Date` directly, but instead have to produce a localised
+ * date/time string. We arbitrarily convert to the New Zealand English format (For example,
+ * '2/04/2024, 11:27:30 am'), since it doesn't seem to be possible to convert directly to a more
+ * easily parsed format like ISO 8601 or RFC 3339.
+ *
+ * @param {string | null} dateTimeString Optional date/time string which can be parsed by the `Date` constructor
+ * @returns {string | undefined} Localised date/time string
+ *
+ */
+function getPacificAucklandYear(dateTimeString: string | null): string | undefined {
+  if (dateTimeString == null) {
+    return undefined;
+  }
+
+  const pacificAucklandDateTimeString = new Date(dateTimeString).toLocaleString('en-NZ', {
+    timeZone: 'Pacific/Auckland',
+  });
+  return pacificAucklandDateTimeString.split(/[,/]/, 4)[2];
 }
 
 /*
@@ -154,9 +177,9 @@ export function formatDate(collection: StacCollection): string {
  * @async
  * @param {string} source
  * @param {StacCollection} collection
- * @returns {Promise<CogTiff>}
+ * @returns {Promise<Tiff>}
  */
-export async function loadFirstTiff(source: string, collection: StacCollection): Promise<CogTiff> {
+export async function loadFirstTiff(source: string, collection: StacCollection): Promise<Tiff> {
   const itemLink = collection.links.find((f) => f.rel === 'item')?.href;
   if (itemLink == null) throw new Error(`No items in collection from ${source}.`);
   const itemPath = new URL(itemLink, source).href;
@@ -170,7 +193,7 @@ export async function loadFirstTiff(source: string, collection: StacCollection):
   return tiff;
 }
 
-export function extractGsd(tiff: CogTiff): number {
+export function extractGsd(tiff: Tiff): number {
   const gsd = tiff.images[0]?.resolution[0];
   if (gsd == null) {
     throw new Error(`Missing resolution tiff tag: ${tiff.source.url}`);
@@ -178,7 +201,7 @@ export function extractGsd(tiff: CogTiff): number {
   return gsd;
 }
 
-export function extractEpsg(tiff: CogTiff): number {
+export function extractEpsg(tiff: Tiff): number {
   const epsg = tiff.images[0]?.epsg;
   if (epsg == null) {
     throw new Error(`Missing epsg tiff tag: ${tiff.source.url}`);